@@ -6,4 +6,4 @@
 @using BlazorSignalR.Test.Client
 @using BlazorSignalR.Test.Client.Shared
 @using BlazorSignalR
-@addTagHelper *, Blazor.Extensions
+@*@using Blazor.Extensions*@
