@@ -21,6 +21,6 @@
 <body>
     <app>@*@(await Html.RenderComponentAsync<App>())*@</app>
 
-    <script src="_framework/components.server.js"></script>
+    <script src="_framework/blazor.server.js"></script>
 </body>
 </html>
